backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
openai==1.55.3
httpx==0.27.2
pandas==2.1.1
numpy==1.25.2
scikit-learn==1.3.1
python-dotenv==1.0.0
python-multipart==0.0.6
PyJWT==2.8.0
cryptography==42.0.5
google-cloud-firestore==2.11.1
google-cloud-storage==2.9.0
openpyxl==3.1.2